+++ src/level/tile/stairs_tile.cpp
@@ -1,6 +1,6 @@
 #include "stairs_tile.hpp"
 
-Stairs_tile::Stairs_tile(int id, bool leads_up) : Tile(id), leads_up{leads_up}
+Stairs_tile::Stairs_tile(uint8_t id, bool leads_up) : Tile(id), leads_up{leads_up}
 {
 }
